feat: download and preprocess

+++ python_code/population/pop_data_download.py
@@ -1,13 +1,14 @@
 import os
 from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
-
+import time
 import numpy as np
 import requests
 import shutil
 from icecream import ic
+from tqdm import tqdm
 
-from my_config import DATA_SRC
+from my_config import DATA_SRC, hd_path_population
 
 # this is needed to avoid the error in the download in MaxOS
 os.environ["no_proxy"] = "*"
@@ -17,28 +18,40 @@ data_population_path.mkdir(parents=True, exist_ok=True)
 tmp_path = data_population_path / "tmp"
 tmp_path.mkdir(parents=True, exist_ok=True)
 base_worldpop_url = "https://data.worldpop.org/GIS/AgeSex_structures/Global_2000_2020/"
-years_range = np.arange(2010, 2018)
-
-# headers = {
-#     'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36',
-#     'X-Forwarded-For': '203.0.113.195',  # Spoofed IP address
-#     'Referer': 'https://google.com',     # Spoofed referer
-#     'Accept-Language': 'en-US,en;q=0.9', # Language preferences
-#     'Accept-Encoding': 'gzip, deflate, br'
-# }
+years_range = np.arange(2000, 2021)
 
 
 def download_file(url, filepath):
     filepath = Path(filepath)
     tmp_filepath = tmp_path / filepath.name
     ic(f"Downloading {filepath}")
-    if not filepath.is_file():
+    if not filepath.is_file() and not tmp_filepath.is_file():
         response = requests.get(url, stream=True)
-        # response = requests.get(url, stream=True, headers=headers)
         response.raise_for_status()
-        with open(tmp_filepath, "wb") as f:
-            for chunk in response.iter_content(chunk_size=8192):
+
+        total_size = int(response.headers.get("content-length", 0))
+        block_size = 8192
+        start_time = time.time()
+        downloaded_size = 0
+
+        with open(tmp_filepath, "wb") as f, tqdm(
+            desc=filepath.name,
+            total=total_size,
+            unit="B",
+            unit_scale=True,
+            unit_divisor=1024,
+        ) as pbar:
+            for chunk in response.iter_content(chunk_size=block_size):
                 f.write(chunk)
+                downloaded_size += len(chunk)
+                pbar.update(len(chunk))
+
+                # Calculate download speed
+                elapsed_time = time.time() - start_time
+                if elapsed_time > 0:
+                    speed = downloaded_size / elapsed_time  # bytes per second
+                    pbar.set_postfix(speed=f"{speed / 1024:.2f} KB/s")
+
         shutil.move(tmp_filepath, filepath)
         ic(f"Downloaded {filepath}")
     else:
@@ -53,7 +66,13 @@ def create_urls_sex_age_years() -> list[tuple[str, str]]:
             for age in [0, 65, 70, 75, 80]:
                 download_url = f"{base_worldpop_url}{year}/0_Mosaicked/global_mosaic_1km/global_{sex}_{age}_{year}_1km.tif"
                 filepath = DATA_SRC / f"population/global_{sex}_{age}_{year}_1km.tif"
-                if not Path(filepath).is_file():
+                tmp_filepath = tmp_path / filepath.name
+                hd_filepath = hd_path_population / filepath.name
+                if (
+                    not filepath.is_file()
+                    and not tmp_filepath.is_file()
+                    and not Path(hd_filepath).is_file()
+                ):
                     urls.append((download_url, filepath))
     return urls
 
@@ -65,21 +84,27 @@ def create_urls_aggregated_years() -> list[tuple[str, str]]:
             f"{base_worldpop_url}{year}/0_Mosaicked/ppp_{year}_1km_Aggregated.tif"
         )
         filepath = DATA_SRC / f"population/ppp_{year}_1km_Aggregated.tif"
-        if not Path(filepath).is_file():
+        tmp_filepath = tmp_path / filepath.name
+        if not filepath.is_file() and not tmp_filepath.is_file():
             urls.append((download_url, filepath))
     return urls
 
 
 if __name__ == "__main__":
     urls = create_urls_sex_age_years()
+    ic(urls)
+    ic(len(urls))
+
+    with ThreadPoolExecutor() as executor:
+        executor.map(lambda p: download_file(*p), urls)
 
     # for url, filepath in urls:
     #     print(url, filepath)
     #     download_file(url, filepath)
 
-    with ThreadPoolExecutor() as executor:
-        executor.map(lambda p: download_file(*p), urls)
-
     # urls = create_urls_aggregated_years()
+    # ic(urls)
+    # ic(len(urls))
+
     # with ThreadPoolExecutor() as executor:
     #     executor.map(lambda p: download_file(*p), urls)
